# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /radars:
    get:
      operationId: listRadars
      tags: [radars]
      summary: List saved Radars
      responses:
        "200":
          description: Saved Radars
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Radar"
    post:
      operationId: createRadar
      tags: [radars]
      summary: Create a saved Radar
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RadarInput"
      responses:
        "201":
          description: Created Radar
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Radar"
  /radars/{radarId}:
    get:
      operationId: getRadar
      tags: [radars]
      summary: Get one Radar
      parameters:
        - $ref: "#/components/parameters/RadarId"
      responses:
        "200":
          description: Radar
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Radar"
        "404":
          description: Radar not found
    patch:
      operationId: updateRadar
      tags: [radars]
      summary: Update or archive a saved Radar
      parameters:
        - $ref: "#/components/parameters/RadarId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RadarUpdate"
      responses:
        "200":
          description: Updated Radar
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Radar"
        "404":
          description: Radar not found
    post:
      operationId: runRadar
      tags: [radars]
      summary: Run a Radar using its configured provider
      parameters:
        - $ref: "#/components/parameters/RadarId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RadarRunInput"
      responses:
        "200":
          description: Radar run result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RadarRun"
        "400":
          description: Radar run failed or missing configuration
          content:
            application/json:
              schema:
                type: object
                required: [error]
                properties:
                  error: { type: string }
                  code: { type: string }
  /leads:
    get:
      operationId: listLeads
      tags: [leads]
      summary: List discovered opportunities
      parameters:
        - name: radarId
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: mode
          in: query
          schema:
            type: string
            enum: [demo, live]
      responses:
        "200":
          description: Discovered leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
  /leads/{leadId}:
    get:
      operationId: getLead
      tags: [leads]
      summary: Get a lead with research evidence
      parameters:
        - $ref: "#/components/parameters/LeadId"
      responses:
        "200":
          description: Lead detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "404":
          description: Lead not found
    patch:
      operationId: updateLead
      tags: [leads]
      summary: Update a lead status or saved state
      parameters:
        - $ref: "#/components/parameters/LeadId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadUpdate"
      responses:
        "200":
          description: Updated lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /outreach:
    get:
      operationId: listOutreach
      tags: [outreach]
      summary: List outreach drafts
      responses:
        "200":
          description: Outreach drafts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Outreach"
    post:
      operationId: createOutreach
      tags: [outreach]
      summary: Create a personalized outreach draft
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OutreachInput"
      responses:
        "201":
          description: Outreach draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Outreach"
  /outreach/{outreachId}:
    patch:
      operationId: updateOutreach
      tags: [outreach]
      summary: Approve, edit, or skip an outreach draft
      parameters:
        - $ref: "#/components/parameters/OutreachId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OutreachUpdate"
      responses:
        "200":
          description: Updated outreach draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Outreach"
...
  /pipeline/{leadId}:
    patch:
      operationId: updatePipelineStage
      tags: [pipeline]
      summary: Move a lead to another pipeline stage
      parameters:
        - $ref: "#/components/parameters/LeadId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PipelineUpdate"
      responses:
        "200":
          description: Updated lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
...
components:
  parameters:
    RadarId:
      name: radarId
      in: path
      required: true
      schema:
        type: string
    LeadId:
      name: leadId
      in: path
      required: true
      schema:
        type: string
    OutreachId:
      name: outreachId
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    Radar:
      type: object
      required: [id, name, description, target, offer, criteria, status, leadCount]
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string }
        target: { type: string }
        offer: { type: string }
        geography: { type: string }
        intent: { type: string }
        criteria:
          type: array
          items: { type: string }
        status: { type: string, enum: [active, paused, archived] }
        leadCount: { type: integer }
    RadarInput:
      type: object
      required: [name, description, target, offer, criteria]
      properties:
        name: { type: string, minLength: 1 }
        description: { type: string }
        target: { type: string }
        offer: { type: string }
        geography: { type: string }
        intent: { type: string }
        criteria:
          type: array
          items: { type: string }
    RadarUpdate:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
        target: { type: string }
        offer: { type: string }
        geography: { type: string }
        intent: { type: string }
        criteria:
          type: array
          items: { type: string }
        status:
          type: string
          enum: [active, paused, archived]
    RadarRunInput:
      type: object
      properties:
        mode:
          type: string
          enum: [demo, live]
          default: demo
    RadarRun:
      type: object
      required: [radar, stages, leadsFound, demoMode]
      properties:
        radar: { $ref: "#/components/schemas/Radar" }
        stages:
          type: array
          items: { type: string }
        leadsFound: { type: integer }
        demoMode: { type: boolean }
        mode:
          type: string
          enum: [demo, live]
        candidatesResearched: { type: integer }
        error: { type: string }
...
    Lead:
      type: object
      required: [id, companyName, description, industry, location, relevance, signals, evidence, opportunity, source, sourceStatus, discoveredAt, radarId, status, saved]
      properties:
        id: { type: string }
        companyName: { type: string }
        website: { type: ["string", "null"] }
        instagram: { type: ["string", "null"] }
        linkedin: { type: ["string", "null"] }
        description: { type: string }
        industry: { type: string }
        location: { type: string }
        founder: { type: ["string", "null"] }
        publicEmail: { type: ["string", "null"] }
        relevance: { type: integer, minimum: 0, maximum: 100 }
        signals:
          type: array
          items: { type: string }
        evidence:
          type: array
          items:
            $ref: "#/components/schemas/Evidence"
        opportunity:
          type: array
          items: { type: string }
        source: { type: string }
        sourceStatus: { type: string, enum: [demo, connected, not_verified] }
        discoveredAt: { type: string }
        radarId: { type: string }
        status:
          type: string
          enum: [discovered, review, shortlisted, outreach_ready, contacted, replied, won, lost, archived]
        saved: { type: boolean }
        contactVerified: { type: boolean }
        fit:
          type: string
          enum: ["HIGH RELEVANCE", "POSSIBLE RELEVANCE", "LOW RELEVANCE"]
        scoreBreakdown:
          type: object
          properties:
            baseScore: { type: integer }
            totalScore: { type: integer }
            fit:
              type: string
              enum: ["HIGH RELEVANCE", "POSSIBLE RELEVANCE", "LOW RELEVANCE"]
            factors:
              type: array
              items:
                type: object
                properties:
                  factor: { type: string }
                  points: { type: integer }
                  reason: { type: string }
        observableSignals:
          type: array
          items:
            type: object
            properties:
              id: { type: string }
              category: { type: string }
              key: { type: string }
              statement: { type: string }
              sourceUrl: { type: string }
              sourceName: { type: string }
              observedAt: { type: string }
              confidence:
                type: string
                enum: [high, medium, low]
              type:
                type: string
                enum: [VERIFIED, INFERRED, SUGGESTED]
              excerpt: { type: string }
        people:
          type: array
          items:
            $ref: "#/components/schemas/Person"
        contactPoints:
          type: array
          items:
            $ref: "#/components/schemas/ContactPoint"
        primaryContact:
          anyOf:
            - $ref: "#/components/schemas/Person"
            - type: "null"
        enrichmentSummary:
          type: string
        opportunityBrief:
          $ref: "#/components/schemas/OpportunityBrief"
    OpportunityBrief:
      type: object
      required:
        - summary
        - whyRelevant
        - opportunity
        - observableFacts
        - inferences
        - contactEvidence
        - recommendedNextAction
        - confidence
        - sources
      properties:
        summary:
          type: string
        whyRelevant:
          type: array
          items:
            type: string
        opportunity:
          type: object
          required:
            - hypothesis
            - explanation
          properties:
            hypothesis:
              type: string
            explanation:
              type: string
        observableFacts:
          type: array
          items:
            $ref: "#/components/schemas/Evidence"
        inferences:
          type: array
          items:
            $ref: "#/components/schemas/Evidence"
        primaryContact:
          anyOf:
            - $ref: "#/components/schemas/Person"
            - type: "null"
        contactEvidence:
          type: object
          required:
            - status
            - details
            - channels
          properties:
            status:
              type: string
              enum: [VERIFIED_DIRECT, SUPPORTED_DIRECT, COMPANY_ONLY, ABSENT]
            details:
              type: string
            primaryEmail:
              type: ["string", "null"]
            channels:
              type: array
              items:
                type: string
        recommendedNextAction:
          type: object
          required:
            - action
            - label
            - reason
          properties:
            action:
              type: string
              enum: [draft_outreach, research_contact, review_evidence, archive_lead]
            label:
              type: string
            reason:
              type: string
        confidence:
          type: string
          enum: [high, medium, low]
        sources:
          type: array
          items:
            type: object
            required:
              - name
              - url
            properties:
              name:
                type: string
              url:
                type: string
              observedAt:
                type: string
              status:
                type: string
    ContactPoint:
      type: object
      required: [id, type, value, scope, verificationStatus, confidence, sourceUrl, sourceName, observedAt, isDirect]
      properties:
        id: { type: string }
        type:
          type: string
          enum: [email, linkedin, twitter, instagram, phone, contact_form]
        value: { type: string }
        scope:
          type: string
          enum: [individual, company]
        verificationStatus:
          type: string
          enum: [VERIFIED, SUPPORTED, NOT_VERIFIED]
        confidence:
          type: string
          enum: [high, medium, low]
        sourceUrl: { type: string }
        sourceName: { type: string }
        observedAt: { type: string }
        isDirect: { type: boolean }
    Person:
      type: object
      required: [id, name, role, roleCategory, verificationStatus, confidence, sourceUrl, sourceName, observedAt, contactPoints]
      properties:
        id: { type: string }
        name: { type: string }
        role: { type: string }
        roleCategory:
          type: string
          enum: [founder, executive, creative_marketing, operations, unspecified]
        verificationStatus:
          type: string
          enum: [VERIFIED, SUPPORTED, SUGGESTED]
        confidence:
          type: string
          enum: [high, medium, low]
        sourceUrl: { type: string }
        sourceName: { type: string }
        observedAt: { type: string }
        excerpt: { type: string }
        selectionReason: { type: string }
        contactPoints:
          type: array
          items:
            $ref: "#/components/schemas/ContactPoint"
    LeadUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [discovered, review, shortlisted, outreach_ready, contacted, replied, won, lost, archived]
        saved: { type: boolean }
    Evidence:
      type: object
      required: [id, statement, sourceName, sourceStatus, observedAt]
      properties:
        id: { type: string }
        statement: { type: string }
        sourceName: { type: string }
        sourceUrl: { type: ["string", "null"] }
        sourceStatus: { type: string, enum: [demo, connected, not_verified] }
        observedAt: { type: string }
        confidence:
          type: string
          enum: [high, medium, low]
        type:
          type: string
          enum: [VERIFIED, INFERRED, SUGGESTED]
        excerpt: { type: string }
    Outreach:
      type: object
      required: [id, leadId, companyName, channel, subject, message, whyThisMessage, status, createdAt, canSend]
      properties:
        id: { type: string }
        leadId: { type: string }
        companyName: { type: string }
        channel: { type: string, enum: [email, instagram, linkedin] }
        subject: { type: string }
        message: { type: string }
        whyThisMessage: { type: string }
        status: { type: string, enum: [draft, approved, skipped] }
        createdAt: { type: string }
        canSend: { type: boolean }
    OutreachInput:
      type: object
      required: [leadId, channel, tone, offer]
      properties:
        leadId: { type: string }
        channel: { type: string, enum: [email, instagram, linkedin] }
        tone: { type: string, enum: [human, casual, direct, professional] }
        offer: { type: string }
        recipientScope:
          type: string
          enum: [person, company]
    OutreachUpdate:
      type: object
      properties:
        subject: { type: string }
        message: { type: string }
        status: { type: string, enum: [draft, approved, skipped] }
...
    PipelineUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [discovered, review, shortlisted, outreach_ready, contacted, replied, won, lost, archived]
...