# lib/api-zod/src/generated/types/outreachInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OutreachInputChannel } from './outreachInputChannel';
import type { OutreachInputRecipientScope } from './outreachInputRecipientScope';
import type { OutreachInputTone } from './outreachInputTone';

export interface OutreachInput {
  leadId: string;
  channel: OutreachInputChannel;
  tone: OutreachInputTone;
  offer: string;
  recipientScope?: OutreachInputRecipientScope;
}