# lib/api-zod/src/generated/types/radar.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RadarStatus } from './radarStatus';

export interface Radar {
  id: string;
  name: string;
  description: string;
  target: string;
  offer: string;
  geography?: string;
  intent?: string;
  criteria: string[];
  status: RadarStatus;
  leadCount: number;
}