# lib/api-zod/src/generated/types/radarInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RadarInput {
  /** @minLength 1 */
  name: string;
  description: string;
  target: string;
  offer: string;
  geography?: string;
  intent?: string;
  criteria: string[];
}