# lib/api-zod/src/generated/types/radarUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RadarUpdateStatus } from './radarUpdateStatus';

export interface RadarUpdate {
  name?: string;
  description?: string;
  target?: string;
  offer?: string;
  geography?: string;
  intent?: string;
  criteria?: string[];
  status?: RadarUpdateStatus;
}